+++ test/integ/util/__init__.py
@@ -2,4 +2,4 @@
 Integration tests for stem.util.* contents.
 """
 
-__all__ = ["conf", "proc", "system"]
+__all__ = ["conf", "connection", "proc", "system"]